Move api cache from types to utils folder
