+++ frontend/src/utils/APICache.ts
